+++ lint/prettier-plugin-smarts/src/index.ts
@@ -201,6 +201,7 @@ const plugin = {
   options,
   defaultOptions: {
     sortJsonKeys: true,
+    importOrder: '',
     vueIndentScriptAndStyle: true,
   },
 };