+++ micro-service/grpc-gateway/src/app.controller.ts
@@ -2,13 +2,13 @@
  * @Author: wangwei wwdqq7@qq.com
  * @Date: 2025-04-21 13:54:27
  * @LastEditors: wangwei wwdqq7@qq.com
- * @LastEditTime: 2025-04-22 18:08:18
+ * @LastEditTime: 2025-04-23 13:57:11
  * @FilePath: /FullStack/micro-service/grpc-gateway/src/app.controller.ts
  * @Description: --
  */
-import { Controller, Post } from '@nestjs/common';
+import { Controller, Post } from '@nestjs/common'
 
-import { AppService } from './app.service';
+import { AppService } from './app.service'
 
 /**
  * 应用控制器
@@ -27,6 +27,6 @@ export class AppController {
 
   @Post()
   getHello(): string {
-    return this.appService.getHello();
+    return this.appService.getHello()
   }
 }